Fix specs to work with Ruby 1.9.3+
The ability to refer to constants in an included module was actually a bug
in Ruby 1.9.2, so the specs were failing on later versions where this bug has
been fixed.

The solution is to define the example blocks inside the module itself, since
that's the only place where the unqualified constants should be visible.

https://github.com/rspec/rspec-core/issues/506
http://bugs.ruby-lang.org/issues/5657
https://twitter.com/dchelimsky/status/185335729738616833